// src/server/flickr/flickr.dtos.ts
export interface FlickrSearchResults {
  photos: {
    page: number;
    pages: number;
    perpage: number;
    total: string;
    photo: {
      id: string;
      owner: string;
      secret: string;
      server: string;
      farm: number;
      title: string;
      ispublic: number;
      isfriend: number;
      isfamily: number;
    }[];
  };
  stat: string;
}